+++ backend/app/__init__.py
@@ -1,8 +1,9 @@
 from flask import Flask
 from flask_cors import CORS
-from app.presentation.controllers import firestore, reproduce, search, home
-import firebase_admin  # type: ignore
-from firebase_admin import credentials, storage # type: ignore
+from app.presentation.controllers import reproduce, search, home
+
+from app.presentation.controllers import cloudinary as cloudinary_controller
+import cloudinary # type: ignore
 import os
 
 from dotenv import load_dotenv
@@ -12,18 +13,19 @@ from dotenv import load_dotenv
 def create_app():
     
     load_dotenv()
-    cred = credentials.Certificate(os.getenv("FIREBASE_SERVICE_ACCOUNT_KEY_PATH"))
-    firebase_admin.initialize_app(cred, {
-    'projectId': os.getenv("FIREBASE_PROJECT_ID"),
-    'storageBucket': os.getenv("FIREBASE_STORAGE_BUCKET")
-    })
+    cloudinary.config( 
+        cloud_name=os.getenv("CLOUDINARY_CLOUD_NAME"), 
+        api_key=os.getenv("CLOUDINARY_API_KEY"), 
+        api_secret=os.getenv("CLOUDINARY_SECRET_KEY"), 
+        secure=True
+    )
     
     app = Flask(__name__)
     CORS(app, resources={r"/*": {"origins": "*"}})
     
-    app.secret_key =  os.getenv("SECRET_KEY")
+    app.secret_key =  os.getenv("FLASK_SECRET_KEY")
 
-    app.register_blueprint(firestore.firestore_bp, url_prefix=f"/{firestore.name}")
+    app.register_blueprint(cloudinary_controller.cloudinary_bp, url_prefix=f"/{cloudinary_controller.name}")
     app.register_blueprint(reproduce.reproduce_bp, url_prefix=f"/{reproduce.name}")
     app.register_blueprint(search.search_bp, url_prefix=f"/{search.name}")
     app.register_blueprint(home.home_bp) # main route does not need url_prefix