+++ backend/app/domain/services/reproduce_manager.py
@@ -1,42 +1,38 @@
-
-
-
-
+from flask import url_for
 from app.presentation.utils.http_response import MyJson
-from firebase_admin import  storage # type: ignore
+import cloudinary  # type: ignore
+from cloudinary.api import NotFound, resource # type: ignore
 
-def reproduce_audio_from_firestore(user_id: str, video_id: str) -> MyJson:
+
+def reproduce_audio_from_cloudinary(user_id: str, video_id: str) -> MyJson:
     try:
-        # Firebase Storage path with user-specific storage
-        file_path = f'users/{user_id}/audio_files/{video_id}.m4a'
+        # Cloudinary Storage path with user-specific storage
+        file_path = f'users/{user_id}/audio_files/{video_id}'
 
-        # Access Firebase Storage
-        bucket = storage.bucket()
-        blob = bucket.blob(file_path)
+        # Check if the file exists on Cloudinary
+        resource_info = resource(file_path, resource_type="video")
+        audio_url = resource_info.get("secure_url")
 
-        if not blob.exists():
-            return MyJson(
-                error="File", 
-                status_code=404, 
-                message="File not found.", 
-                data=[]
-            )
-
-        # Generate signed URL (valid for 1 hour)
-        audio_url = blob.generate_signed_url(expiration=3600)
-
-        # Return signed URL
+        
+        
         return MyJson(
             error="No Error", 
             status_code=200, 
-            message="Generated signed URL, valid for 1 hour.", 
+            message="Audio file found successfully.", 
             data=audio_url
-        )
-
+            )
+    except NotFound:
+        
+        return MyJson(
+            error="FileNotFound",
+            status_code=404,
+            message="The requested audio file does not exist on Cloudinary.",
+            data=url_for('cloudinary.upload_music', user_id=user_id, video_id=video_id)
+    )
     except Exception as e:
         return MyJson(
-            error=str(e), 
-            status_code=500, 
-            message="An unknown error occurred.", 
-            data=[]
+            error="ServerError",
+            status_code=500,
+            message=str(e),
+            data=""
         )