+++ frontend/lib/data/remote/interface_auth.dart
@@ -1,8 +1,10 @@
+import 'package:firebase_auth/firebase_auth.dart';
 
 abstract class AuthInterface {
   Future<String?> loginWithGoogle();
   Future<String?> loginWithFacebook();
   Future<void> logout();
+  User? currentUser();
 }