+++ frontend/lib/domain/music_player.dart
@@ -1,5 +1,6 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:frontend/data/remote/firebase_auth_adpter.dart';
 import 'package:frontend/data/utils/paths.dart';
 import 'package:frontend/domain/services/logs.dart';
 import 'package:frontend/presentation/assets/l10n/generated/l10n.dart';
@@ -18,6 +19,7 @@ class MusicPlayer extends StatefulWidget {
 class _MusicPlayerState extends State<MusicPlayer> {
   late AudioPlayer _audioPlayer;
   final AppLogger logger = AppLogger();
+  final FirebaseAuthAdapter _firebaseAuthAdapter = FirebaseAuthAdapter();
   
 
   @override
@@ -29,12 +31,16 @@ class _MusicPlayerState extends State<MusicPlayer> {
   }
 
   Future<void> _loadMusic(String videoId) async {
-    final url = 'http://${LocalApiPath.baseUrl}${LocalApiPath.routes.reproduceAudio()}$videoId';
+
+
+    String userId = _firebaseAuthAdapter.currentUser()!.uid;
+    
+    final url = 'http://${LocalApiPath.baseUrl}${LocalApiPath.routes.reproduceAudio()}$userId/$videoId';
     try {
       await _audioPlayer.setUrl(url);
       await _audioPlayer.play();
     } catch (e) {
-      logger.debug("Error loading audio: $e");
+      return ;
     }
   }
 
@@ -159,37 +165,55 @@ class PlayerControls extends StatelessWidget {
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
         IconButton(
-          icon: Icon(Icons.skip_previous),
+          icon: Icon(Icons.skip_previous,size: 40),
           onPressed: onPrevious,
         ),
         StreamBuilder<PlayerState>(
-          stream: audioPlayer.playerStateStream,
-          builder: (context, snapshot) {
-            
-            if (snapshot.connectionState == ConnectionState.waiting) {
-                return const Align(
-                  alignment: Alignment.center,
-                  child: SizedBox(
-                    width: 20,
-                    height: 20,
-                    child: CircularProgressIndicator()),
-                );
-        }
-        if (snapshot.hasError || !snapshot.hasData) {
-          return Icon(Icons.error_outline, size: 40);
-        }
-            final state = snapshot.data;
-            final isPlaying = state?.playing ?? false;
-            return IconButton(
-              icon: Icon(isPlaying ? Icons.pause : Icons.play_arrow),
-              onPressed: () {
-                isPlaying ? audioPlayer.pause() : audioPlayer.play();
-              },
-            );
-          },
+  stream: audioPlayer.playerStateStream,
+  builder: (context, snapshot) {
+    if (snapshot.connectionState == ConnectionState.waiting) {
+      return const Align(
+        alignment: Alignment.center,
+        child: SizedBox(
+          width: 20,
+          height: 20,
+          child: CircularProgressIndicator(),
         ),
+      );
+    }
+
+    if (snapshot.hasError || !snapshot.hasData) {
+      return const Icon(Icons.error_outline, size: 40);
+    }
+
+    final state = snapshot.data;
+    final isPlaying = state?.playing ?? false;
+    final isLoading = state?.processingState == ProcessingState.loading ||
+        state?.processingState == ProcessingState.buffering;
+
+    if (isLoading) {
+      return const SizedBox(
+        width: 40,
+        height: 40,
+        child: CircularProgressIndicator(),
+      );
+    }
+
+    return IconButton(
+      icon: Icon(
+        isPlaying ? Icons.pause : Icons.play_arrow,
+        size: 40,
+        ),
+      onPressed: () {
+        isPlaying ? audioPlayer.pause() : audioPlayer.play();
+      },
+    );
+  },
+),
         IconButton(
-          icon: Icon(Icons.skip_next),
+          icon: Icon(Icons.skip_next,
+        size: 40
+        ),
           onPressed: onNext,
         ),
       ],