+++ CHANGELOG.md
@@ -38,8 +38,11 @@ All notable changes to this project will be documented in this file.
 
 ### <!-- 1 -->🐛 Bug Fixes
 
+-   **frontend**:  Loading process icon running as wait to load music ([57bfe95](57bfe95275d31b2ec22c24596d647ec150197543))
+-   **app**:  Fixed the reproduce process to play the song completely ([400a12c](400a12c492443cc891cd398869bd235ad8b8d1cc))
+-   **backend**:  Fixed blueprint name on __init__ for cloudinary ([39d4ba6](39d4ba6fabb7b7e264d88e7456b214aa421d6ea5))
+-   Force-applied stash and overwrote local changes ([3e4077c](3e4077c487bd1382a78c624f54769ef2b2b44333))
 -   **frontend**:  Fixed the function that search videos from yt that consume th api ([d3bb873](d3bb8730b6150ff5e3f95973161dd10cc02c7df6))
--   **frontend**:  Fixed the function that search videos from yt that consume th api ([c244f50](c244f50dd417c7d9875bfce7b7808057859ca180))
 -   **backend**:  Application run is available for test communication with frontend ([c6b88ce](c6b88cea72ad7ae152380f052033220f23079d8d))
 -   **backend**:  Start refactoring the api to load the music correctly ([d71aa19](d71aa19f7ee9dfac9bab038a7de86231d0a4a858))
 -   App.py file was empty, so was full filled with the right code ([7dc3c94](7dc3c94acce4981fe882dceab6ac3bca84f9d0a8))